+++ CHANGELOG.md
@@ -480,6 +480,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Recursion
 
+## 0102 - Binary Tree Level Order Traversal
+
+- Recursion
+
 ## 0415 - Add strings
 
 - Full adder